Fix IndexOutOfRangeException when including in collection with dropped index, reorganized issue tests
